TTO-207 investigate discrepancy in hathifiles full vs. incremental
`cutoff` variable used in `generate_hathifile.rb` was excluding HTIDs that arguably changed due
to changes in the catalog record. This patch gets rid of that variable and treats all HTIDs
as having been affected by the record change (thus including them in the resulting hathifile)
regardless of their 974(d) update date.

Tests should be refactored at some point but I consider it out of scope for a TTO issue.